# back-end/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "GoGym API",
        "description": "API REST para gerenciamento de academias, personal trainers e alunos. Permite registro de usuários (personal e aluno), autenticação via token (Sanctum), verificação de e-mail, CRUD completo de personais e alunos, e sistema de convites para vinculação entre personal e aluno.",
        "contact": {
            "name": "Gustavo",
            "email": "nao-responda@ugustavodev.com.br"
        },
        "license": {
            "name": "Proprietária",
            "url": "https://ugustavodev.com.br"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:84/api",
            "description": "Servidor de Desenvolvimento"
        }
    ],
    "paths": {
...
        "/email/verificar/{id}/{hash}": {
            "get": {
                "tags": [
                    "Verificação de E-mail"
                ],
                "summary": "Verificar endereço de e-mail",
                "description": "Valida o link de verificação enviado por e-mail. A URL contém uma assinatura temporária que expira em 60 minutos. Este endpoint é acessado diretamente pelo link do e-mail.",
                "operationId": "verificarEmail",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID do usuário",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    },
                    {
                        "name": "hash",
                        "in": "path",
                        "description": "Hash SHA1 do e-mail do usuário",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "expires",
                        "in": "query",
                        "description": "Timestamp de expiração da URL assinada",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "signature",
                        "in": "query",
                        "description": "Assinatura criptográfica da URL",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "E-mail verificado com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "mensagem": {
                                            "type": "string",
                                            "example": "E-mail verificado com sucesso!"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Link inválido, expirado ou hash incorreto",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "mensagem": {
                                            "type": "string",
                                            "example": "Link de verificação inválido ou expirado."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Usuário não encontrado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErroNaoEncontrado"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/email/reenviar": {
            "post": {
                "tags": [
                    "Verificação de E-mail"
                ],
                "summary": "Reenviar e-mail de verificação",
                "description": "Reenvia o e-mail de verificação para o usuário autenticado. Requer autenticação, mas não requer e-mail verificado.",
                "operationId": "reenviarVerificacao",
                "responses": {
                    "200": {
                        "description": "E-mail de verificação reenviado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "mensagem": {
                                            "type": "string",
                                            "example": "E-mail de verificação reenviado."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Não autenticado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErroNaoAutenticado"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "E-mail já verificado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "mensagem": {
                                            "type": "string",
                                            "example": "E-mail já verificado."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        }
    },
    "components": {
        "schemas": {
...
            "ErroNaoAutenticado": {
                "properties": {
                    "message": {
                        "type": "string",
                        "example": "Unauthenticated."
                    }
                },
                "type": "object"
            },
...
            "ErroNaoEncontrado": {
                "properties": {
                    "message": {
                        "type": "string",
                        "example": "Recurso não encontrado."
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Insira o token retornado no login/registro. Exemplo: {token}",
                "scheme": "bearer",
                "bearerFormat": "Sanctum"
            }
        }
    },
...
}